# Brain_API_Specification.yaml
openapi: 3.1.0
info:
  title: Brain API
  version: 0.3.0
  description: |
    The Brain API exposes six layers as HTTP endpoints plus one MCP surface.
    Raw (evidence), Ledger (normalized financial truth), Wiki (memory),
    Policy (governance), Agent (action orchestration), Audit (proof).

    v0.2.0 introduces the Normalized Ledger layer. The Wiki API is preserved
    but its role narrows to memory rendering and Q&A — financial truth lives
    in the Ledger. The Execution endpoints are renamed to Agent endpoints;
    legacy /execution/* paths remain shipped with deprecation headers for
    the duration of the v0.3 transition.

    v0.3.0 aligns the spec to the canonical public docs at
    https://docs.brain.fi. Material changes land in small commits — see
    docs/sdk-audit.md for the reconciliation queue. This release aligns
    the error envelope to the docs shape (`{ error: { code, message,
    details, trace_id, docs_url } }`); the v0.2 flat shape is replaced.

    All endpoints are tenant-scoped. Tenant identity is derived from the
    authenticated principal. The principal's tenant is never passed in URLs
    except where explicitly required for cross-tenant operations by platform
    administrators.

    Base URL (production): https://api.brain.fi/v1
    Base URL (sandbox):    https://api.brain.dev/v1
    Protocol: HTTPS only. HTTP requests are rejected at the edge.
    Content type: application/json unless otherwise specified.
    Character encoding: UTF-8.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/siwx/challenge:
    post:
      tags: [Admin]
      security: []  # public — no bearer token required
      summary: Request a SIWX challenge nonce
      description: |
        Issues a fresh nonce and `session_id` the agent should echo back
        on `POST /auth/siwx`. The nonce is bound to a 5-minute TTL and
        is single-use. Optional path; clients may also provide their own
        nonce inside the SIWX message.
      operationId: siwxChallenge
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_address:
                  type: string
                  pattern: '^0x[a-fA-F0-9]{40}$'
                  description: |
                    Optional hint — the server may use it to pre-fetch
                    the agent's registration record so the verify call
                    is faster.
      responses:
        '200':
          description: Challenge issued
          content:
            application/json:
              schema:
                type: object
                required: [nonce, session_id, domain]
                properties:
                  nonce: { type: string, minLength: 8 }
                  session_id: { type: string }
                  domain:
                    type: string
                    description: |
                      The EIP-4361 `domain` claim the agent must include
                      in its signed message. Defaults to `api.brain.fi`.
  /auth/siwx:
    post:
      tags: [Admin]
      security: []  # public — caller is unauthenticated until verify succeeds
      summary: Sign in as an external agent via SIWX
      description: |
        Verifies an EIP-4361-formatted message signed with the agent's
        registered key, checks the agent is `active` in the local
        `agents` table and that its `scope_hash` matches the on-chain
        `BrainMCPAgentRegistry` record, then issues an `agent_token`
        JWT (1-hour TTL, refreshable by re-signing).

        v0.3 ship verifies the signature and shape; the on-chain
        scope-hash check is stubbed until the registry contract
        integration lands (see PLAN-FIRST #14b).
      operationId: siwxVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, signature]
              properties:
                message:
                  type: string
                  description: EIP-4361 SIWX message in plain text.
                signature:
                  type: string
                  pattern: '^0x[a-fA-F0-9]+$'
                session_id:
                  type: string
                  description: |
                    The session id returned by /auth/siwx/challenge.
                    Required when the message's nonce was not pre-shared
                    out-of-band.
      responses:
        '200':
          description: Verified — agent_token issued
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_in, principal]
                properties:
                  access_token:
                    type: string
                    description: |
                      Brain agent JWT. Use as
                      `Authorization: Bearer <access_token>` on
                      subsequent calls.
                  token_type:
                    type: string
                    enum: [Bearer]
                  expires_in:
                    type: integer
                    description: Seconds until token expiry.
                  principal:
                    type: object
                    properties:
                      id: { type: string }
                      type: { type: string, enum: [agent] }
                      tenantId: { type: string }
                      scopes:
                        type: array
                        items: { type: string }
        '401':
          description: |
            Signature verification or agent registration check failed.
            Maps to one of:
              * `auth_siwx_invalid` — signature did not recover to the
                expected address, or the nonce was expired/already used.
              * `agent_not_found` — recovered address is not registered.
              * `agent_inactive` — agent is registered but not in
                `active` state.
              * `scope_hash_mismatch` — JWT scope_hash would not match
                the on-chain registry value.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  #############################################################################
  # SOURCES — source-connection lifecycle (PLAN-FIRST #12)
  #
  # Adapter-facing routes that the Brain Console and SDK call to manage
  # ingestion connections. Each row in `raw_sources` represents one
  # active connection to a provider (Plaid bank, Stripe processor, etc.).
  # Sources feed `/raw/ingest` behind the scenes — webhooks land in
  # `/raw/webhooks/{provider}`, polled adapters call `/raw/ingest`
  # internally.
  #
  # Source: https://docs.brain.fi/api-reference/sources-api.
  #############################################################################
...
  /agents/{agent_id}:
    get:
      tags: [Agent]
      summary: Agent detail
      operationId: getAgent
      parameters:
        - name: agent_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Agent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token issued by Brain's auth service.
        Contains tenant_id and scopes. See Engineering Standards
        section 2 for the full auth model.
...
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        Canonical Brain error envelope. Every non-2xx response wraps its
        body in this shape. Source of truth:
        https://docs.brain.fi/resources/errors (canonical wire codes) and
        https://docs.brain.fi/api-reference/overview (envelope shape).

        Engineering Standards §4.1 previously specified the same nested
        shape under the field name `request_id`; the docs publish it as
        `trace_id` and the docs win — see docs/sdk-audit.md conflict E.
      properties:
        error:
          type: object
          required: [code, message, trace_id, docs_url]
          properties:
            code:
              type: string
              description: |
                Machine-readable error code. Stable: once shipped a code
                is never renamed (Standards §4.3 "additions only"). See
                services/api/src/shared/errors.ts for the registry.
            message:
              type: string
              description: Human-readable error message (English, untranslated).
            details:
              type: object
              additionalProperties: true
              description: |
                Optional structured payload. Schema is code-specific —
                e.g. `policy_denied` carries `{rule_id, …}`, `gate_*`
                carries `{check, …}`.
            trace_id:
              type: string
              description: |
                Correlation id for support. Per-request, surfaced in
                logs, metrics, and traces. Pass to support to retrieve
                the full request timeline.
              example: trc_8f3a92...
            docs_url:
              type: string
              format: uri
              description: |
                Direct link to the docs page for this error code, of the
                form https://docs.brain.fi/errors/{code}.
              example: https://docs.brain.fi/errors/policy_denied
      example:
        error:
          code: policy_denied
          message: Counterparty not approved
          details:
            counterparty_id: cp_x
            policy_version: 3
          trace_id: trc_8f3a92abc
          docs_url: https://docs.brain.fi/errors/policy_denied

    #
    # Sources (PLAN-FIRST #12) — adapter connections that feed the Raw
    # layer. The 8-value MVP set per docs/sdk-audit.md decision K2.
    # Adding a new value to `SourceType` is a non-breaking enum widening
    # — extends as adapters ship (Architecture §3.1, "Other
    # source-specific adapters … are post-MVP").
    #
...
    Agent:
      type: object
      required: [id, kind, role, display_name, state]
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [internal, external]
        role:
          type: string
          enum: [reconciliation, payment, anomaly, partner]
        display_name: { type: string }
        scope_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
          nullable: true
        onchain_address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
          nullable: true
        state:
          type: string
          enum: [pending_onchain, active, revoked, failed]
        registered_tx: { type: string, nullable: true }
        registered_at: { type: string, format: date-time, nullable: true }

    #
    # MCP — JSON-RPC 2.0 envelopes for the /agents/mcp surface
    #
...